+++ packages/ssmenv-cli/src/commands/env/list.ts
@@ -1,6 +1,7 @@
 import { Command } from '@oclif/command';
+import { Environment } from 'ssmenv';
 
-import { getDirectEnvironment } from '../../config/fs';
+import { getAwsConfig } from '../../config/fs/helpers';
 import { make as makeExample } from '../../example';
 
 export class EnvList extends Command {
@@ -22,26 +23,15 @@ export class EnvList extends Command {
   ];
 
   async run() {
-    const proxy = await getDirectEnvironment();
-    const request = {
-      Path: '/',
-      Recursive: true,
+    const configFile = await getAwsConfig();
+    const awsConfig = await configFile.read();
+    const instanceConfig = {
+      accessKeyId: awsConfig.accessKeyId,
+      apiVersion: '2014-11-06',
+      region: 'us-east-1',
+      secretAccessKey: awsConfig.secretAccessKey,
     };
-    const results = await proxy.getParametersByPath(request);
-    const parameters = results.Parameters || [];
-    parameters
-      .map(param => {
-        const lastSlash = param.Name!.lastIndexOf('/');
-        return param.Name!.slice(0, lastSlash);
-      })
-      .reduce((paths: string[], next: string) => {
-        if (paths.includes(next)) {
-          return paths;
-        } else {
-          return [...paths, next];
-        }
-      }, [])
-      .sort()
-      .forEach(path => this.log(path));
+    const results: string[] = await Environment.listAll(instanceConfig);
+    results.forEach(path => this.log(path));
   }
 }